Implemented support to edit deck in DecksPage
Now it's possible for the user to edit a deck by clicking in its menu
and in 'edit' option. The deck component itself is changed from
'display' to 'edit', and the deck is auto focused in its 'name' field.

Two options can be done:

* if the content of the deck is changed and 'Enter' is pressed, then
the deck is updated with the new content.
* if 'Esc' is pressed the update is canceled and the component back to
'display' mode again.