Changed flux structure to fit with pages features
Because of the defined pages structure, using react-router as engine to
manage their flow, the actual flux archicteture doesn't work very well.
The listeners used to listen to changes in decks, cards etc, made
updates to stores in a global app space, regardless of the page being
used. This feature looks good at first sight because it looks like the
application will maintain a central repository of stores that will be
used by all pages. But this funcionality doesn't work well with the
pages mainly because of the following reason:

When the user access the deck page, directly, by its URL, then the
application will be started at that point, and the mechanisms of
listeners will be started, so starting the loading of decks, cards
etc. However, the deck page was expecting a already loaded deck to
display it, and because they're loading yet, the page will probably
crash with an error.

The solution for this problem was to start listeners in each page,
accordingly to their needs. Each page is now responsible to perfom
the actions needed to start the listeners of data it will need in
its content.