Improved application architecture
This change was made mainly to clean up the code but also to improve
the login process in each page. The prior architecture was driven by
flux actions as the responsibles for signing in/out the user. This
approach made the code less readable and consequently harder to
maitain.

In this new architecture approach, in each page where the user needs
to be signed in, the process of listening to 'authStateChange' is set
up at those pages.