Reset all the application architecture
All the application architecture was remade to fit into the project
desires, which are cleanliness, easy maintence and others related
stuff.

All the project was deleted and created again.
The current stage of the project has a simple implementation of the
decks page, which a box component to show, edit, and delete deck. The
application also has implemented some features like add deck, edit deck
and partially delete deck (it must also delete nested cards). The
application is using firebase as its back-end.