Development checkpoint commit
This commit is just for checking point development progress. Because
one of the already implemented features is going to be replaced in a
test, this commit was created to allow back in history if the new
feature not work well.

In this commit we already we the following features implemented:
* Decks handling (add, edit, remove)
* Components for decks and cards (this will be replaced).