Update Deck Smallbox component to improve UI
The old way Deck SmallBox worked was using buttons to confirm and
cancel changes. This way worked great for this deck component but
the same technique wouldn't work as good as it for the Card component.
In the card component, we realized that a more dynamic aproach would be
better; an approach where the used just type things and his focus/blur
actions will determine which decision to make: save card, cancel
changing etc. For keeping consistency within all our UI, we decided to
change the way the Deck component works, and make its UI to work the
same approach as the Card component does.