RESET ALL PROJECT STRUCTURE
The whole project was reseted because from this point its architecture
was intented to be more clear for who is starting developing a project
from scratch.

We are no longer using create-react-app for scaffolding neither other
alternative tools. We are going to use webpack configurations done by
our hands; for each needed feature we will configure our enviroment.

This way of doing the project will be more usefull for learning
purposes for us and whoever is using this project as a way for
studying react, flux, and all the frameworks and tools this project is
using.