Add webpack support
We created a very basic webpack configuration just for testing if it's
working correctly. We also added index.js and src/test.js files for
testing if the module bundling works properly. These files must be
removed in the next iterations.

Git ignore file was changed to reflect the node_modules and build
folder.