Update build process to contain simple HTML page
In this point we have a mechanism to build the project with
an index.html page. This process is done by the build process which now
contains a build script which builds the application using the webpack
config.

To test the application, do the following:

> npm run build
> cd build/
> open index.html

A hello world page should be shown.