Add support to firebase
The applicatio now has support to firebase back-end engine through its
API. In the firebase folder we have to configurations, one for
development and other for production; the appropriate configuration is
loaded accordingly to the node current environment.

Now, in the initial page, we exhibit a simple signing message to
demostrate firebase support. This message shows the progress of signing
in of a anonymous user.