Update build process to improve modules loading
Before this change the application build process was configured to
handle CSS (SCSS) files in the common configuration file. This was
acceptable because either for development as production those kind of
files need to be process. But from this point, each configuration, for
each type of enviroment, will process the CSS (SCSS) files accordingly
to their needs.