Add eslint support
The airbnb style guide is used as entry point and defaults. Now, the
build process contains a lint option to run the linter with 'fix'
option enabled. It also contains a call to the lint script before the
build script; this enables linting process to be guaranteed before
releasing buit versions.

Some fixes needed to be applied to reflect the new lint configurations.