Add pages mapping definition
Now the application contains the definition of the pages it will use to
interact with the user. It contains mapping for the following pages:
Decks, one deck, study, and 404 page. The root page redirects to the
decks page.

The pages haven't any content yet, but just its routes were defined.
The application is ready for navigation between its already defined
pages.