Update build process to include css modules
Every file which ends with '.m.css', '.module.css', '.m.scss' and
'.module.scss' will be considered a CSS module by the build
process. The main page contains an example component which contains
a yellow background, a height of 40px and a brown border, to tests if
the modules configuration is working right.