Update build process to improve pre hooks
Now the pre-commit and pre-push actions will be executed including
a lint without fix. This is needed because when linting with fix
options duration a pre-commit action, this doesn't cause the commit to
fail, but fix files but not in the stage area.

With this update, now when the commit is going to be executed, a non
fix lint will execute to stop the action if there are code erros and
invalid syntax.