Important project structure refactor
All the application structure was refactored to apply to a folder
structure driven by feature instead of layers. Now the src/ root
folders describe, each one, a feature of the application. Some other
folders like firebase, bulma and common are also in that src/ root
folder because it's like a high level feature of the application too.