Fix logo bug in tablets devices
The logo was being edge cutted and the 'Flashcards' words wasn't being
shown correctly in some tablet devices. The solution was to redesign
the logo image and instead of creating a large logo image, only a small
logo with only the figure part of the logo was used, and for the text
part of the logo, a simple text tag was used.