Add support to firebase along with redux
It was added support to anonymous authentication. The firebase is
connected to get cards and decks from it. It has two firebase
configuration files, one for development, other for production. It was
also added support pages to exhibition in case of loading while in
signing in state, and other for authentication problem. The main page
was separated from root.