Refactor app architecture to improve data handling
Now the application is ready for data reading from firebase services.
It has implemented an independent system of handling decks with the
firebase, and it was done using Proxy pattern for handling backend
connected data as simple POJO objects.

The whole components structure was also refactored to add a new
technique for working with components stylization. It was removed
support to Sass, CSS modules and even vanilla CSS. The project is
intended to use styled-components, but this is in a test stage yet and
it's not decided which stylization technique will be used.