Update application architecture
The application was changed to have a better support to offline
operations. It now has some indicators of data not updated, updating
etc. The application data model was refactor to reflect this need. And
all pages were refactored too.
A more simple services API was implemented to ease understanding and
clean the code.